scripts: build-combined-fwbundle: use pre-built per-board fwbundles
In order to reduce churn in CI and re-use previously built artifacts,
adjust `./scripts/build-combined-fwbundle.sh` to interpret an
environment variable `BUNDLE_TEMP_PREFIX`, which (if it exists) can
be used to prefix per-board firmware bundles.

If `BUNDLE_TEMP_PREFIX` is in the environment, then a check is made
to see if `${BUNDLE_TEMP_PREFIX}/$REV/update.fwbundle` exists.

If it does exist, then it is used instead of re-building all targets
for a given board revision `$REV`.

Testing Done:
```
BUNDLE_TEMP_PREFIX="$PWD/build-" ./scripts/build-combined-fwbundle.sh
Using existing BUNDLE_TEMP_PREFIX: build-
Building release 18.1.99 / pack 18.1.99.0
Using pre-built build-p100/update.fwbundle
Using pre-built build-p100a/update.fwbundle
Using pre-built build-p150a/update.fwbundle
Using pre-built build-p150b/update.fwbundle
Using pre-built build-p150c/update.fwbundle
Using pre-built build-p300a/update.fwbundle
Using pre-built build-p300b/update.fwbundle
Using pre-built build-p300c/update.fwbundle
Creating fw_pack-18.1.99.0.fwbundle
Wrote fwbundle to fw_pack-18.1.99.0.fwbundle

echo "still works without the environment variable specified"
./scripts/build-combined-fwbundle.sh
Created TEMP_DIR
Building release 18.1.99 / pack 18.1.99.0
Building tt_blackhole@p100/tt_blackhole/smc
...
```

+++ scripts/build-combined-fwbundle.sh
@@ -5,12 +5,16 @@
 
 set -e
 
-TEMP_DIR="$(mktemp -d)"
-
-echo "Created TEMP_DIR"
+if [ -z "$BUNDLE_TEMP_PREFIX" ]; then
+  TEMP_DIR="$(mktemp -d)/"
+  echo "Created TEMP_DIR"
+else
+  echo "Using existing BUNDLE_TEMP_PREFIX: $BUNDLE_TEMP_PREFIX"
+  TEMP_DIR="$BUNDLE_TEMP_PREFIX"
+fi
 
 cleanup() {
-    if [ -d "$TEMP_DIR" ]; then
+    if [ -z "$BUNDLE_TEMP_PREFIX" ] && [ -d "$TEMP_DIR" ]; then
         rm -Rf "$TEMP_DIR"
         echo "Removed TEMP_DIR"
     fi
@@ -18,15 +22,17 @@ cleanup() {
 trap cleanup EXIT
 
 mPATH="$(dirname "$0")"
+TTZP_BASE="$(dirname "$mPATH")"
 # put tt_boot_fs.py in the path (use it to combine .fwbundle files)
 PATH="$mPATH:$PATH"
 
+# TODO: read boards / board revs from a YAML file
 BOARD_REVS="p100 p100a p150a p150b p150c p300a p300b p300c"
 
-MAJOR="$(grep "^VERSION_MAJOR" VERSION | awk '{print $3}')"
-MINOR="$(grep "^VERSION_MINOR" VERSION | awk '{print $3}')"
-PATCH="$(grep "^PATCHLEVEL" VERSION | awk '{print $3}')"
-EXTRAVERSION="$(grep "^EXTRAVERSION" VERSION | awk '{print $3}')"
+MAJOR="$(grep "^VERSION_MAJOR" "$TTZP_BASE/VERSION" | awk '{print $3}')"
+MINOR="$(grep "^VERSION_MINOR" "$TTZP_BASE/VERSION" | awk '{print $3}')"
+PATCH="$(grep "^PATCHLEVEL" "$TTZP_BASE/VERSION" | awk '{print $3}')"
+EXTRAVERSION="$(grep "^EXTRAVERSION" "$TTZP_BASE/VERSION" | awk '{print $3}')"
 
 if [ -z "$MAJOR" ] || [ -z "$MINOR" ] || [ -z "$PATCH" ]; then
   echo "required version info missing: MAJOR=$MAJOR MINOR=$MINOR PATCH=$PATCH"
@@ -48,15 +54,23 @@ echo "Building release $RELEASE / pack $PRELEASE"
 for REV in $BOARD_REVS; do
   BOARD="tt_blackhole@$REV/tt_blackhole/smc"
 
+  if [ -n "$BUNDLE_TEMP_PREFIX" ]; then
+    if [ -f "${TEMP_DIR}${REV}/update.fwbundle" ]; then
+      echo "Using pre-built ${TEMP_DIR}${REV}/update.fwbundle"
+      continue
+    fi
+    echo "Warning: pre-built ${TEMP_DIR}${REV}/update.fwbundle not found"
+  fi
+
   echo "Building $BOARD"
-  west build -d "$TEMP_DIR/$REV" --sysbuild -p -b "$BOARD" app/smc >/dev/null 2>&1
+  west build -d "${TEMP_DIR}${REV}" --sysbuild -p -b "$BOARD" app/smc >/dev/null 2>&1
 done
 
 echo "Creating fw_pack-$PRELEASE.fwbundle"
 # construct arguments..
 ARGS=""
 for REV in $BOARD_REVS; do
-  ARGS="$ARGS -c $TEMP_DIR/$REV/update.fwbundle"
+  ARGS="$ARGS -c ${TEMP_DIR}${REV}/update.fwbundle"
 done
 
 # shellcheck disable=SC2086
